Refactor and improve code readability across multiple files
- Updated macro definitions in Sprite.h for better formatting.
- Enhanced memory handling in Spriteset.c by casting size to size_t in memcpy.
- Improved struct formatting in Spriteset.h for consistency.
- Refactored CreateBlendTables and DeleteBlendTables in Tables.c for clarity and consistency in error handling.
- Cleaned up Test.c by standardizing indentation and formatting.
- Enhanced Tilemap.c with consistent formatting and improved error handling.
- Updated Tilemap.h for consistent struct formatting.
- Refactored TLN_LoadWorld and TLN_ReleaseWorld in World.c for improved readability and structure.
- Ensured consistent error handling and formatting across all modified files.

+++ src/Spriteset.c
@@ -110,7 +110,7 @@ bool TLN_SetSpritesetData(TLN_Spriteset spriteset, int entry, TLN_SpriteData con
         uint8_t const *src = (uint8_t *)pixels;
         uint8_t *dst = TLN_GetBitmapPtr(spriteset->bitmap, data->x, data->y);
         for (int c = 0; c < data->h; c++) {
-            memcpy(dst, src, data->w);
+            memcpy(dst, src, (size_t)data->w);
             src += pitch;
             dst += spriteset->bitmap->pitch;
         }
